Separate routes and models
